Editor: intIn clamps in Long, converts at the boundary
integralValueOf returns Long (the delay field needs the width), so
coercing it against Int bounds had no matching overload - the CI
compile failure. The bounds widen to Long for the clamp and narrow
back at the return, where the value is already in Int's range.

+++ app/src/main/java/dev/eclipse/ssh/ui/editor/EditorPrefs.kt
@@ -127,7 +127,9 @@ object EditorPrefsCodec {
      */
     private fun JSONObject.intIn(key: String, fallback: Int, min: Int, max: Int): Int {
         val value = integralValueOf(key) ?: return fallback
-        return value.coerceIn(min, max).toInt()
+        // Long throughout, Int only at the boundary: integralValueOf speaks Long (the delay
+        // field needs it), and the Int bounds widen without loss — Int.MAX_VALUE fits a Long.
+        return value.coerceIn(min.toLong(), max.toLong()).toInt()
     }
 
     /**
